fix(bugs): state transition guards, initBugTicketsTable in getBugTicketStatus, extract buildBackUrl
- resolveBugTicket: add AND status = 'open' guard so already-resolved tickets cannot be re-resolved
- archiveBugTicket: add AND status != 'archived' guard to prevent no-op double archives
- getBugTicketStatus: add initBugTicketsTable() call for consistency with other read functions
- Extract buildBackUrl into _helpers.ts to eliminate verbatim duplication in resolve/archive endpoints
